Criação do hook useSearch
